pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asr-jetson"
version = "0.1.0"
description = "Lightweight ASR Pipeline with Diarization for Jetson Orin Nano"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = {text = "MIT"}
authors = [{ name = "Xavier Bertrand", email = "xavier.bertrand@example.com" }]
keywords = ["asr", "speech-recognition", "diarization", "jetson", "whisper", "nemo"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
  "numpy<2.0",
  "scikit-learn",
  "tqdm",
  "pydub",
  "soundfile",
  "librosa==0.10.2.post1",
  "ffmpeg-python",
  "faster-whisper",
  "onnx>=1.14",
  "jiwer<4",
  "marshmallow",
  "pyloudnorm",
  "resampy",
  "wrapt",
  "braceexpand",
  "editdistance",
  "wget",
  "tensorboard",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "pyyaml",
  "huggingface_hub>=0.23.0,<0.24.0",
  "transformers==4.41.2",
  "ctranslate2>=4.4.0",
  # --- PyTorch auto multi-arch (sans extras) ---
  # Windows x86_64 CUDA 12.4
  "torch==2.4.0+cu124; platform_system == 'Windows' and platform_machine == 'AMD64'",
  "torchaudio==2.4.0+cu124; platform_system == 'Windows' and platform_machine == 'AMD64'",
  # Linux x86_64 (WSL2 inclus) CUDA 12.4
  "torch==2.4.0+cu124; platform_system == 'Linux' and platform_machine == 'x86_64'",
  "torchaudio==2.4.0+cu124; platform_system == 'Linux' and platform_machine == 'x86_64'",
  # Jetson Linux aarch64 (cuDNN/CUDA fournis par JetPack ; pas de suffixe +cu124)
  "torch>=2.0.0; platform_system == 'Linux' and platform_machine == 'aarch64'",
  "torchaudio>=2.0.0; platform_system == 'Linux' and platform_machine == 'aarch64'",
  "language-tool-python>=2.9.4",
  "speechbrain>=1.0.3",
  "pyannote-audio>=3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "pytest-xdist>=3.3.0",
  "pytest-asyncio>=0.21.0",
  "black>=23.0.0",
  "ruff>=0.1.0",
  "mypy>=1.5.0",
  "pre-commit>=3.4.0",
]
api = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",
]
# Installe TensorRT/PyCUDA seulement sur Jetson (Linux aarch64)
tensorrt = [
  "tensorrt; platform_system == 'Linux' and platform_machine == 'aarch64'",
  "pycuda; platform_system == 'Linux' and platform_machine == 'aarch64'",
]
# PyAV : version compatible multi-plateformes
media = [
  "av>=13,<15",
]

# PyTorch pour Windows avec CUDA 12.4
gpu-win = [
  "torch==2.4.0+cu124; platform_system == 'Windows'",
  "torchaudio==2.4.0+cu124; platform_system == 'Windows'",
]

# PyTorch pour Linux x86_64 (WSL2, Desktop Linux) avec CUDA 12.4
gpu-linux = [
  "torch==2.4.0+cu124; platform_system == 'Linux' and platform_machine == 'x86_64'",
  "torchaudio==2.4.0+cu124; platform_system == 'Linux' and platform_machine == 'x86_64'",
]

# PyTorch pour Jetson ARM64 (depuis PyTorch officiel ou NVIDIA)
# Sur Jetson, PyTorch est généralement pré-installé ou installé via pip standard
gpu-jetson = [
  "torch>=2.0.0; platform_system == 'Linux' and platform_machine == 'aarch64'",
  "torchaudio>=2.0.0; platform_system == 'Linux' and platform_machine == 'aarch64'",
]

# NeMo ASR (Linux uniquement : desktop Linux, WSL2, Jetson)
nemo = [
  "nemo_toolkit[asr]>=1.21.0,<2.0; platform_system == 'Linux'",
]

[project.urls]
Homepage = "https://github.com/XavBertrand/ASR_jetson"
Repository = "https://github.com/XavBertrand/ASR_jetson"
Issues = "https://github.com/XavBertrand/ASR_jetson/issues"

[project.scripts]
asr-pipeline = "asr_jetson.pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["asr_jetson*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "-v",
  "--strict-markers",
  "--tb=short",
  "--cov=src/asr_jetson",
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-report=xml",
]
markers = [
  "integration: integration tests (slow, may require models)",
  "unit: fast unit tests",
  "gpu: tests requiring GPU",
  "slow: slow tests (> 5 seconds)",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::PendingDeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]
exclude = [".git", "__pycache__", ".venv", "venv", "build", "dist"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ["tests/", "build/"]

[tool.coverage.run]
source = ["src/asr_jetson"]
omit = [
  "*/tests/*",
  "*/test_*.py",
  "*/__pycache__/*",
  "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu124"]
override-dependencies = [
  "huggingface-hub>=0.23.0,<0.24.0",
]

[tool.uv.extra-build-dependencies]
youtokentome = ["Cython"]
